[clients] Remove dummyLogger
